tests reformat extracted BaseRecord merge two mysql files
