+++ tests/unit/DynamicActiveRecordTest.php
@@ -2,10 +2,12 @@
 
 namespace tests\unit;
 
+use tests\unit\data\BaseRecord;
 use tests\unit\data\dar\Supplier;
 use Yii;
 use spinitron\dynamicAr\DynamicActiveRecord;
 use spinitron\dynamicAr\DynamicActiveQuery;
+use yiiunit\framework\db\ActiveRecordTest;
 use yiiunit\framework\db\DatabaseTestCase;
 use yiiunit\TestCase;
 use tests\unit\data\dar\Product;
@@ -13,36 +15,21 @@ use yii\db\Connection;
 use yii\db\Query;
 use yii\db\ActiveQuery;
 
-class DynamicActiveRecordTest extends DatabaseTestCase
+class DynamicActiveRecordTest extends ActiveRecordTest
 {
-    /*
-     * link(,,$extraColumns)
-     *
-     */
     /** @var Connection */
     protected $db;
 
     const BINARY_STRING = "\x00\x01\x02\x03\x04\x05\x06\x07\x08\x09\x0a\x0b\x0c\x0d\x0e\x0f\x10\x11\x12\x13\x14\x15\x16\x17\x18\x19\x1a\x1b\x1c\x1d\x1e\x1f\x20\x21\x22\x23\x24\x25\x26\x27\x28\x29\x2a\x2b\x2c\x2d\x2e\x2f\x30\x31\x32\x33\x34\x35\x36\x37\x38\x39\x3a\x3b\x3c\x3d\x3e\x3f\x40\x41\x42\x43\x44\x45\x46\x47\x48\x49\x4a\x4b\x4c\x4d\x4e\x4f\x50\x51\x52\x53\x54\x55\x56\x57\x58\x59\x5a\x5b\x5c\x5d\x5e\x5f\x60\x61\x62\x63\x64\x65\x66\x67\x68\x69\x6a\x6b\x6c\x6d\x6e\x6f\x70\x71\x72\x73\x74\x75\x76\x77\x78\x79\x7a\x7b\x7c\x7d\x7e\x7f\x80\x81\x82\x83\x84\x85\x86\x87\x88\x89\x8a\x8b\x8c\x8d\x8e\x8f\x90\x91\x92\x93\x94\x95\x96\x97\x98\x99\x9a\x9b\x9c\x9d\x9e\x9f\xa0\xa1\xa2\xa3\xa4\xa5\xa6\xa7\xa8\xa9\xaa\xab\xac\xad\xae\xaf\xb0\xb1\xb2\xb3\xb4\xb5\xb6\xb7\xb8\xb9\xba\xbb\xbc\xbd\xbe\xbf\xc0\xc1\xc2\xc3\xc4\xc5\xc6\xc7\xc8\xc9\xca\xcb\xcc\xcd\xce\xcf\xd0\xd1\xd2\xd3\xd4\xd5\xd6\xd7\xd8\xd9\xda\xdb\xdc\xdd\xde\xdf\xe0\xe1\xe2\xe3\xe4\xe5\xe6\xe7\xe8\xe9\xea\xeb\xec\xed\xee\xef\xf0\xf1\xf2\xf3\xf4\xf5\xf6\xf7\xf8\xf9\xfa\xfb\xfc\xfd\xfe\xff";
 
-    protected function resetFixture()
-    {
-        $fixture = __DIR__ . '/maria10.sql';
-        $this->db->createCommand(file_get_contents($fixture))->execute();
-    }
-
     protected static $resetFixture = true;
 
     protected function setUp()
     {
-		static::$params = require(__DIR__ . '/data/config.php');
-		parent::setUp();
-		$this->db = $this->getConnection();
+        static::$params = require(__DIR__ . '/data/config.php');
+        parent::setUp();
 
-        Product::$db = $this->db;
-        Supplier::$db = $this->db;
-        if (self::$resetFixture) {
-            $this->resetFixture();
-        }
+        $this->db = BaseRecord::$db = $this->getConnection(self::$resetFixture);
     }
 
     public function testSetAttribute()
@@ -89,8 +76,8 @@ class DynamicActiveRecordTest extends DatabaseTestCase
         $p->setAttribute('array.c.c.c', 7);
         $this->assertEquals($array, $p->array);
 
-        $array['c']['c'] = [9,8,7,6];
-        $p->setAttribute('array.c.c', [9,8,7,6]);
+        $array['c']['c'] = [9, 8, 7, 6];
+        $p->setAttribute('array.c.c', [9, 8, 7, 6]);
         $this->assertEquals($array, $p->array);
     }
 
@@ -129,7 +116,7 @@ class DynamicActiveRecordTest extends DatabaseTestCase
 
     public function testAsArray()
     {
-        /** @var Product */
+        /** @var Product $product */
         $product = Product::find()->one();
         $expect = [
             'name' => 'product1',
@@ -203,6 +190,7 @@ class DynamicActiveRecordTest extends DatabaseTestCase
                 "_x\xE1" => false,
             ],
         ];
+
         return array_merge($tests);
     }
 
@@ -341,8 +329,129 @@ class DynamicActiveRecordTest extends DatabaseTestCase
         );
         unset($product);
     }
+
+//    public function testFindAsArray()
+//    {
+//        /* @var $customerClass \yii\db\ActiveRecordInterface */
+//        $customerClass = $this->getCustomerClass();
+//
+//        // asArray
+//        $customer = $customerClass::find()->where(['id' => 2])->asArray()->one();
+//        // since address is commented in the schema - it should be saved as a dynamic column
+//        $this->assertEquals([
+//            'id' => 2,
+//            'email' => 'user2@example.com',
+//            'name' => 'user2',
+//            'status' => 1,
+//            'profile_id' => null,
+//            'dynamic_columns' => '{"address":"address2"}',
+//        ], $customer);
+//
+//        // find all asArray
+//        $customers = $customerClass::find()->asArray()->all();
+//        $this->assertEquals(3, count($customers));
+//
+//        // since address is commented in the schema - it should be saved as a dynamic column
+//        $this->assertArrayHasKey('dynamic_columns', $customers[0]);
+//        $this->assertArrayNotHasKey('address', $customers[0]);
+//
+//        $this->assertArrayHasKey('id', $customers[0]);
+//        $this->assertArrayHasKey('name', $customers[0]);
+//        $this->assertArrayHasKey('email', $customers[0]);
+//        $this->assertArrayHasKey('status', $customers[0]);
+//
+//        // since address is commented in the schema - it should be saved as a dynamic column
+//        $this->assertArrayHasKey('dynamic_columns', $customers[1]);
+//        $this->assertArrayNotHasKey('address', $customers[1]);
+//
+//        $this->assertArrayHasKey('id', $customers[1]);
+//        $this->assertArrayHasKey('name', $customers[1]);
+//        $this->assertArrayHasKey('email', $customers[1]);
+//        $this->assertArrayHasKey('status', $customers[1]);
+//
+//        // since address is commented in the schema - it should be saved as a dynamic column
+//        $this->assertArrayHasKey('dynamic_columns', $customers[2]);
+//        $this->assertArrayNotHasKey('address', $customers[2]);
+//
+//        $this->assertArrayHasKey('id', $customers[2]);
+//        $this->assertArrayHasKey('name', $customers[2]);
+//        $this->assertArrayHasKey('email', $customers[2]);
+//        $this->assertArrayHasKey('status', $customers[2]);
+//    }
+//
+//    public function testFindIndexByAsArray()
+//    {
+//        /* @var $customerClass \yii\db\ActiveRecordInterface */
+//        $customerClass = $this->getCustomerClass();
+//
+//        /* @var $this TestCase */
+//        // indexBy + asArray
+//        $customers = $customerClass::find()->asArray()->indexBy('name')->all();
+//        $this->assertEquals(3, count($customers));
+//        // since address is commented in the schema - it should be saved as a dynamic column
+//        $this->assertArrayHasKey('dynamic_columns', $customers['user1']);
+//        $this->assertArrayNotHasKey('address', $customers['user1']);
+//
+//        $this->assertArrayHasKey('id', $customers['user1']);
+//        $this->assertArrayHasKey('name', $customers['user1']);
+//        $this->assertArrayHasKey('email', $customers['user1']);
+//        $this->assertArrayHasKey('status', $customers['user1']);
+//
+//        // since address is commented in the schema - it should be saved as a dynamic column
+//        $this->assertArrayHasKey('dynamic_columns', $customers['user2']);
+//        $this->assertArrayNotHasKey('address', $customers['user2']);
+//
+//        $this->assertArrayHasKey('id', $customers['user2']);
+//        $this->assertArrayHasKey('name', $customers['user2']);
+//        $this->assertArrayHasKey('email', $customers['user2']);
+//        $this->assertArrayHasKey('status', $customers['user2']);
+//
+//        // since address is commented in the schema - it should be saved as a dynamic column
+//        $this->assertArrayHasKey('dynamic_columns', $customers['user3']);
+//        $this->assertArrayNotHasKey('address', $customers['user3']);
+//
+//        $this->assertArrayHasKey('id', $customers['user3']);
+//        $this->assertArrayHasKey('name', $customers['user3']);
+//        $this->assertArrayHasKey('email', $customers['user3']);
+//        $this->assertArrayHasKey('status', $customers['user3']);
+//
+//        // indexBy callable + asArray
+//        $customers = $customerClass::find()->indexBy(function ($customer) {
+//            return $customer['id'] . '-' . $customer['name'];
+//        })->asArray()->all();
+//        $this->assertEquals(3, count($customers));
+//        // since address is commented in the schema - it should be saved as a dynamic column
+//        $this->assertArrayHasKey('dynamic_columns', $customers['1-user1']);
+//        $this->assertArrayNotHasKey('address', $customers['1-user1']);
+//
+//        $this->assertArrayHasKey('id', $customers['1-user1']);
+//        $this->assertArrayHasKey('name', $customers['1-user1']);
+//        $this->assertArrayHasKey('email', $customers['1-user1']);
+//        $this->assertArrayHasKey('status', $customers['1-user1']);
+//
+//        // since address is commented in the schema - it should be saved as a dynamic column
+//        $this->assertArrayHasKey('dynamic_columns', $customers['1-user1']);
+//        $this->assertArrayNotHasKey('address', $customers['1-user1']);
+//
+//        $this->assertArrayHasKey('id', $customers['2-user2']);
+//        $this->assertArrayHasKey('name', $customers['2-user2']);
+//        $this->assertArrayHasKey('email', $customers['2-user2']);
+//        $this->assertArrayHasKey('status', $customers['2-user2']);
+//
+//        // since address is commented in the schema - it should be saved as a dynamic column
+//        $this->assertArrayHasKey('dynamic_columns', $customers['1-user1']);
+//        $this->assertArrayNotHasKey('address', $customers['1-user1']);
+//
+//        $this->assertArrayHasKey('id', $customers['3-user3']);
+//        $this->assertArrayHasKey('name', $customers['3-user3']);
+//        $this->assertArrayHasKey('email', $customers['3-user3']);
+//        $this->assertArrayHasKey('status', $customers['3-user3']);
+//    }
+
     // todo need to test all the active query uses and relations
     // the way to do it, i think, is to follow yii2's ActiveRecordTest
     // as a guide. everything in there should be adapted to dynamic-ar if it makes sense.
     // https://github.com/yiisoft/yii2/blob/master/tests/unit/framework/db/ActiveRecordTest.php
+
+    // todo test creating and that there is 'CREATE COLUMN' fragment in sql
 }
